+++ modules/documentation/docs/changelog.md
@@ -2,11 +2,20 @@
 
 ## Next Release
 
+## 0.1.15-beta.15
+
+- \[router\] Fix bug in auto-rebalance
+## 0.1.15-beta.14
+
+- \[router\] Add auto-rebalance service for Matic deposit
+
+## 0.1.15-beta.13
 
 - \[ops\] Add persistent storage for Grafana
 - \[contracts\] Add `transactionHash` to withdrawal commitment
 - \[store\] Add `transferDefinition` to filter options on `getTransfers`
 - \[store\] Add `getWithdrawalCommitmentByTransactionHash` method
+- \[types\] Add `createdAt`/`resolvedAt` to `FullTransferState.meta` + make transfer meta types generic
 
 ## 0.1.15-beta.12